Fix scroll view dropping grid shadow on popover
This commit also does a little refactoring - namely, methods related to
HoverViewController and DefinitionViewController are now an extension for EditViewController.
